+++ src/naive.rs
@@ -8,7 +8,7 @@ use std::option::{Option, Some, None};
 pub fn rank<T: Access<bool>>(v: &T, bit: bool, n: int) -> int {
     let mut accum = 0;
     for i in range(0, n) {
-        if v.get(i) == bit {
+        if v.get(i as uint) == bit {
             accum += 1;
         }
     }
@@ -18,7 +18,7 @@ pub fn rank<T: Access<bool>>(v: &T, bit: bool, n: int) -> int {
 pub fn select<T: Access<bool> + Collection>(v: &T, bit: bool, n: int) -> Option<int> {
     let mut n = n;
     for i in range(0, v.len()) {
-        if v.get(i as int) == bit {
+        if v.get(i) == bit {
             if n == 0 {
                 return Some(i as int);
             }
